industry_helper runs on both backends; v0.9.56
Twenty-four statements moved from `?` + tuple to named binds, four hand-built
`IN (?,?,?)` strings became expanding bindparams, and two `executemany` calls
became SQLAlchemy executemany over dicts — guarded, because an empty list is not
a no-op there the way it was before: with no rows to infer the parameter shape
from, SQLAlchemy raises rather than doing nothing.

The assertions that guard it are the ones from the previous commit, unchanged.
That was the point of writing them first: a test written after a rewrite can
only describe what the rewrite did, where these had to survive it. All that
moved underneath them is the fixture, and `tests/test_industry_helper_rigs.py`
became `tests/test_industry_helper_on_postgres.py`.

**A live defect, found by running those tests against Postgres for the first
time: eight failures there, none on SQLite.**

`station_rigs.location_id` was declared `Integer`. SQLite's INTEGER is a
variable-width 64-bit type and treats the declaration as advisory, so it has
stored structure ids happily for the life of the project. A Postgres INTEGER is
exactly 32 bits, and an Upwell structure id is around 1.03e12 — so on Postgres
that table accepted NPC stations and rejected every player-owned structure,
which is the only kind that can carry rigs at all. Migration 0009 widens it.

It is one column of a class. ESI declares character, corporation, location,
item, order and contract ids as int64 and the schema still calls several of them
`Integer`; the list is in the worklist rather than swept in here, because a
widening ALTER on a live table is not something to do by pattern match. The
reusable form: SQLite ignores the width in a declaration and Postgres enforces
it, so every integer column here is unverified until something pushes a
real-world-sized value through it on Postgres.

Seven mutations, and the two that matter are asymmetric. Neutering migration
0009 fails **8 Postgres tests and 0 SQLite ones**; removing the dialect guard
from the schema shim fails **9 and 0**. That asymmetry is the only evidence that
the cross-backend half tests a backend difference rather than running the same
assertions twice.

One of those mutations was wrong before it was right, which is worth recording.
Reverting the column type **in the declaration** failed nothing anywhere, and
the first reading was "the assertion is decorative". It was not: the Postgres
fixture builds its tables from `upgrade_to_head`, so the type there comes from
the migration, and the declaration is read only by `create_all` and by
`test_the_migrations_match_the_declaration` — which does fail, on its own, when
the declaration drifts. Two artifacts, two claims, two mutations. Check which
one the test actually reads before concluding it guards nothing.

That guard also caught the first version of migration 0009 skipping SQLite
entirely. Semantically that changes nothing there, but it left the declaration
and the history disagreeing on the backend the test runs.

`test_the_computed_percentage_stacks_multiplicatively` is marked `sqlite_only`,
and the marker is a to-do with a name. Four functions here ask
`location_resolver` for a security multiplier whenever a station has rigs
fitted, and it still builds `?` placeholders. So this module is on the portable
layer without being Postgres-clean end to end. Deleting that marker is the
definition of done for the next slice.

The boundary count went 4 → 5 rather than staying flat as predicted. Three
crossings left `margins.py` and `reactions_helper.py`; three appeared inside
`industry_helper`; and a fifth appeared because three handlers in
`routers/locations.py` had to move onto `connect()` too. All five now point at
`location_resolver`, so the number falls to zero in one slice rather than in
three. Those three handlers also stopped leaking on failure — each closed with a
bare `conn.close()` that an exception skips, and one wraps its whole body in
`except Exception`, so the leak had no symptom at all.

Two fixtures and two tests were building schema or crossing boundaries that no
longer exist. Each failed loudly at the call site rather than quietly doing the
wrong thing, which is the failure mode worth having.

Dropped one assertion deliberately: an earlier test pinned that the same rig in
two slots counts twice. Nobody fits the same rig twice, so it was pinning a case
that cannot occur — and if it cannot occur, deduplicating and not deduplicating
are the same total, and the test protected nothing. Replaced with two different
rigs both counting, which is the real loadout.

788 tests green, 2 skipped.

+++ app/db/schema.py
@@ -376,7 +376,14 @@ facility_tax_cache = Table(
 
 station_rigs = Table(
     "station_rigs", metadata,
-    Column("location_id", Integer, primary_key=True, autoincrement=False),
+    # BigInteger, not Integer: an Upwell structure id is around 1.03e12, which
+    # is fine in SQLite (whose INTEGER is 64-bit whatever the declaration says)
+    # and out of range for a Postgres INTEGER. Declared narrow, this table
+    # accepts NPC stations and rejects every player-owned structure — which is
+    # the only kind that can have rigs at all. Found by running the rig tests
+    # against Postgres; see the `int64 ids` note in docs/step-4-worklist.md,
+    # because this is one column of a wider class.
+    Column("location_id", BigInteger, primary_key=True, autoincrement=False),
     Column("me_bonus_pct", Float, nullable=False, server_default="0"),
     Column("updated_at", BigInteger, nullable=False),
     Column("structure_type", Text),
